chore(masonry): move new brick components to src/components
