+++ modules/masonry/src/components/Brick/Brick.tsx
