+++ modules/masonry/src/components/Path/Path.tsx
